Compilo, y ejecuto configuracion en nueva libreria externa YAML, en el servidor, todo ok

+++ common/protocol/game_action/game_action.cpp
@@ -86,6 +86,6 @@ void GameAction::setPlayerId(uint32_t id) {
     playerId = id;
 }
 
-uint8_t GameAction::getPlayerId() {
+uint32_t GameAction::getPlayerId() { // Antes estaba puesto en uint8_t
     return playerId;
 }
