+++ server/game/body/chell.h
@@ -11,11 +11,11 @@
 class Chell: public Body {
 private:
     void createBody(float32 xPos, float32 yPos) override;
-    uint8_t playerId;
+    uint32_t playerId;
     void move(float32 xSpeed, float32 ySpeed);
 public:
-    Chell(b2World &world, float32 xPos, float32 yPos, uint8_t playerId);
-    uint8_t getPlayerId();
+    Chell(b2World &world, float32 xPos, float32 yPos, uint32_t playerId);
+    uint32_t getPlayerId();
     ~Chell() override;
     void moveLeft();
     void moveRight();
